_smtp-agent/requirements.txt
firebase-admin==6.2.0
python-dotenv==1.0.0
Flask==3.0.2
Flask-HTTPAuth==4.8.0
